fix(flutter): support analyzer 13 test runtime

+++ plugins/design-system-guardian/adapters/flutter/lib/src/config/adapter_config.dart
@@ -546,7 +546,7 @@ final class GuardianAdapterConfigRepository {
     }
     return ConfigBinding.valid(
       GuardianAdapterConfig(
-        schemaVersion: schemaVersion as int,
+        schemaVersion: schemaVersion,
         profileId: profileId,
         policyDigest: policyDigest,
         snapshotId: snapshotId,
